+++ HISTORY.md
@@ -1,6 +1,13 @@
 ## v18.0.3
-* Locale Update: street address prefixes in `fr` locale.
-* Added `Faker.GenerateForever` that returns `IEnumerable<T>` with unlimited items when iterated over.
+* Locale Updates - 
+* `fr`: new street address prefixes.
+* `fa`: new street addresses.
+* `pl`: removed 2008 value from city.
+* `en`: new gender first names
+* New Dutch (Belgium) `nl_BE` locale.
+* New Romanian `ro` locale.
+* Added `f.Finance.RoutingNumber` - Generates an ABA routing number with valid check digit.
+* Added `Faker.GenerateForever` that returns `IEnumerable<T>` with unlimited generated items when iterated over.
 
 ## v18.0.2
 * Issue 86: Removed diacritic mark/accents (á, í, ó, ú, etc) from generated email addresses and user names.